requirements.txt
python-jobspy>=1.1
httpx[http2]>=0.27
pydantic>=2.0
fake-useragent>=1.4
pandas>=2.0
python-dotenv>=1.0
openpyxl>=3.1
